fixes: Processing garbage works as intended - updated class and method signatures - removed wrong .toFloat cast - proper garbage processing and logging - added separate clocks for sources, sinks and collectors

+++ src/main/scala/mission/impossibl/bots/sink/Models.scala
@@ -8,13 +8,13 @@ import java.util.UUID
 
 final case class Garbage(amount: Int)
 
-final case class Instance(id: UUID, location: (Int, Int), storageCapacity: Float, orchestrator: ActorRef[GarbageOrchestrator.Command])
+final case class Instance(id: UUID, location: (Int, Int), storageCapacity: Int, orchestrator: ActorRef[GarbageOrchestrator.Command])
 
 //todo waste type?
-final case class GarbagePacketRecord(wasteSourceId: UUID, wasteType: Int = 0, wasteMass: Float)
+final case class GarbagePacketRecord(wasteSourceId: UUID, wasteType: Int = 0, wasteMass: Int)
 
-final case class GarbagePacket(records: List[GarbagePacketRecord], totalMass: Float)
+final case class GarbagePacket(records: List[GarbagePacketRecord], totalMass: Int)
 
-final case class State(id: UUID, efficiency: Float, garbagePackets: List[GarbagePacket] = List.empty, reservedSpace: List[Reservation] = List.empty)
+final case class State(id: UUID, efficiency: Int, garbagePackets: List[GarbagePacket] = List.empty, reservedSpace: List[Reservation] = List.empty)
 
-final case class Reservation(auctionId: UUID, garbageCollectorId: UUID, wasteMass: Float, timeout: Cancellable)
+final case class Reservation(auctionId: UUID, garbageCollectorId: UUID, wasteMass: Int, timeout: Cancellable)
